Use base metaclass for source and sink contexts
