+++ montblanc/impl/rime/tensorflow/sources/source_context.py
@@ -18,13 +18,8 @@
 # You should have received a copy of the GNU General Public License
 # along with this program; if not, see <http://www.gnu.org/licenses/>.
 
-import functools
-import inspect
-import types
-
-from hypercube import HyperCube
-
-from montblanc.impl.rime.tensorflow.context_help import context_help
+from ..context_help import context_help
+from ..hypercube_proxy_metaclass import HypercubeProxyMetaClass
 
 class _setter_property(object):
     def __init__(self, func, doc=None):
@@ -33,46 +28,6 @@ class _setter_property(object):
 
     def __set__(self, obj, value):
         return self.func(obj, value)
-
-class SourceContextMetaClass(type):
-    """ MetaClass for Sink Contexts """
-    def __init__(cls, name, bases, dct):
-        """ Proxy public methods on the HyperCube """
-        def public_member_predicate(m):
-            return inspect.ismethod(m) and not m.__name__.startswith('_')
-
-        hc_members = inspect.getmembers(HyperCube, public_member_predicate)
-        sc_members = inspect.getmembers(cls, public_member_predicate)
-
-        intersect = set.intersection(
-            set(m[0] for m in hc_members),
-            set(m[0] for m in sc_members))
-
-        if len(intersect) > 0:
-            raise ValueError("Proxying methods failed on class '{c}'. "
-                "The following members '{m}' conflicted with class '{hc}'."
-                    .format(c=cls.__name__, m=list(intersect), hc=HyperCube.__name__))
-
-        def wrap_cube_method(name, method):
-            def _proxy(self, *args, **kwargs):
-                return getattr(self._cube, name)(*args, **kwargs)
-
-            wrap = functools.update_wrapper(_proxy, method)
-            spec = inspect.getargspec(method)
-            fmt_args = inspect.formatargspec(formatvalue=lambda v: '=_default', *spec)
-            call_args = inspect.formatargspec(formatvalue=lambda v: '', *spec)
-
-            # wrap.__doc__ = (
-            #     'def {}{}:\n'
-            #     '\t""" {} """\n'
-            #     '\treturn _proxy{}').format(name, fmt_args, method.__doc__, call_args)
-
-            return wrap
-
-        for name, method in hc_members:
-            setattr(cls, name, wrap_cube_method(name, method.__func__))
-
-        super(SourceContextMetaClass, cls).__init__(name, bases, dct)
 
 class SourceContext(object):
     """
@@ -114,9 +69,9 @@ class SourceContext(object):
 
     """
     __slots__ = ('_cube', '_cfg', '_name', '_shape', '_dtype',
-        '_iter_args', '_array_schema', '_cube_attributes')
+        '_iter_args', '_array_schema')
 
-    __metaclass__ = SourceContextMetaClass
+    __metaclass__ = HypercubeProxyMetaClass
 
     def __init__(self, name, cube, slvr_cfg, iter_args,
                 array_schema, shape, dtype):
